+++ ae-sparse.py
@@ -78,7 +78,7 @@ for fname in os.listdir("TestSongs/"):
   print(encoded_songs[count][0])
   print(encoded_songs[count][1])
   plt.scatter(encoded_songs[count][0], encoded_songs[count][1], s=700,
-              c=(encoded_songs[count][1]/10.0,0,1-encoded_songs[count][1]/10.0),
+              c=(int(encoded_songs[count][1]/10.0),0,int(1-encoded_songs[count][1]/10.0)),
               marker=r"$ {} $".format(fname[:2]), edgecolors='none')
   count += 1
 plt.show()
